convert coin-flip-example to ts
