+++ frontend/src/components/Navigation.tsx
@@ -1,0 +1,50 @@
+import Image from "next/image";
+import Link from "next/link";
+import { useEffect, useState } from "react";
+import { useWalletSelector } from "@near-wallet-selector/react-hook";
+
+
+export const Navigation = () => {
+  const { signedAccountId, signIn, signOut } = useWalletSelector();
+
+  const [action, setAction] = useState<(() => void) | null>(null);
+  const [label, setLabel] = useState<string>("Loading...");
+
+  useEffect(() => {
+    if (signedAccountId) {
+      setAction(() => signOut);
+      setLabel(`Logout ${signedAccountId}`);
+    } else {
+      setAction(() => signIn);
+      setLabel("Login");
+    }
+  }, [signedAccountId, signIn, signOut]);
+
+  return (
+    <nav className="navbar navbar-expand-lg">
+      <div className="container-fluid">
+        <Link href="/" passHref legacyBehavior>
+          <a>
+            <Image
+              priority
+              src="/near-logo.svg"
+              alt="NEAR"
+              width={30}
+              height={24}
+              className="d-inline-block align-text-top"
+            />
+          </a>
+        </Link>
+        <div className="navbar-nav pt-1">
+          <button
+            className="btn btn-secondary"
+            onClick={action ?? undefined}
+            disabled={!action}
+          >
+            {label}
+          </button>
+        </div>
+      </div>
+    </nav>
+  );
+};
