+++ frontend/src/global.d.ts
@@ -1,0 +1,6 @@
+declare module "*.png" {
+  const value: string;
+  export default value;
+  
+}
+declare module "*.module.css";
